Bring simulator to front on run/debug (#611).
Fixes: #611.
